backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.4

# Cache (Valkey — NOT Redis)
valkey==6.0.2

# HTTP client (for vLLM, ML service, Groq fallback)
httpx==0.27.2
tenacity==9.0.0

# Object storage
minio==7.2.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2

# Logging
structlog==24.4.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0

# WebSockets
websockets==13.1
